Pipeline/requirements.txt
torch
facenet-pytorch
Pillow
opencv-python
ultralytics
google-generativeai
gemini-api
